Allowing passing RESP
This allows to run the clean script using watch

+++ clean-checkpoints.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
-echo "The next checkpoint folders will be deleted (all but 5 more recent):"
+echo "The next chekpoint folders will be deleted (all but 5 more recent):"
 ls -dt models/${TAG-bertsification}-*/checkpoint*/ | awk 'NR>5'
-read -p "Do you want to continue? (y/n) " RESP
+if [ -z "$RESP" ]; then
+  read -p "Do you want to continue? (y/n) " RESP
+fi
 if [ "$RESP" = "y" ]; then
   echo "Deleting..."
   rm -rf `ls -dt models/${TAG-bertsification}-*/checkpoint*/ | awk 'NR>5'`
